split dbs to separate files
