+++ src/server/db/index.js
@@ -1,0 +1,3 @@
+import { InMemory } from './inmemory';
+
+export { InMemory };
